Added Count() builder (#5)
* Added Count() builder

* Rework dataclass construction to support command specific columns

In order to support the Count command, the dynamic data classes need to
extract the attributes from the command.

The ColumnDefinitions are moved to the SQLCommand and are overridden
by Select (which supports joins) and Count (which adds a custom column count)

Dataclass construction itself is no longer recursive, but is built from the
column definitions at once.

* Configuration of github action

* github action installs poetry

* github action installs poetry with curl

* Fix for poetry install

* Moved python out of container and exposed service container port

* Added python 3.12 to matrix

* Updated pylint to >3.0.0

* Fixed linting for python 3.12

* Add pypi.yaml workflow